+++ NEWS.rst
@@ -13,11 +13,15 @@ Version 0.6.2 - 2014-05-09
   * NEW: support for BODY, 3DSOLID and REGION entity, you get the ACIS data
   * CHANGE: Spline() - removed context managers fit_points(), control_points(), knot_values() and weights() and added a
     general context_manager edit_data(), similar to Mesh.edit_data() - unified API
-  * CHANGE: MText.buffer() -> MText.edit_data() - unified API (MText.buffer() still exist as alias)
+  * CHANGE: MText.buffer() -> MText.edit_data() - unified API (MText.buffer() still exists as alias)
   * CHANGE: refactored internal structure - only two DXF factories remaining:
 
     - LegacyDXFFactory() for AC1009 (DXF12) drawings
     - ModernDXFFactory() for newer DXF versions except DXF13/14.
+
+  * BUGFIX: LWPolyline.get_rstrip_point() removed also x- and y-coords if zero
+  * BUGFIX: opens DXF12 files without handles again
+  * BUGFIX: opens DXF12 files with HEADER section but without $ACADVER set
 
 Version 0.6.1 - 2014-05-02